Add incident generator (#3)
* Add structure and style for incident

* Add and remove incidents with buttons

* Add random incident message

Solve warnings in eslint.
Add error in random incident.

+++ src/components/SecondaryButton.jsx
@@ -1,23 +1,45 @@
-import React, { useContext } from "react";
+import React, { useContext, useEffect, useState } from "react";
 import { FaTrash } from "react-icons/fa";
 import AppContext from "../context/AppContext";
 import "../styles/Buttons.css";
 
-const SecondaryButton = () => {
-  const { state, deleteAllVisits } = useContext(AppContext);
+const SecondaryButton = ({ type, children }) => {
+  const { state, deleteAllVisits, deleteIncident } = useContext(AppContext);
+  const [disabled, setDisabled] = useState(true);
 
   const handleClick = () => {
-    // eslint-disable-next-line no-restricted-globals
-    const isExecuted = confirm("¿Estas segura de eliminar todas las visitas?");
-    if (isExecuted) deleteAllVisits();
+    if (type === "visit") {
+      // eslint-disable-next-line no-restricted-globals
+      const isExecuted = confirm(
+        "¿Estas segura de eliminar todas las visitas?"
+      );
+      if (isExecuted) deleteAllVisits();
+    }
+    if (type === "incident") {
+      // eslint-disable-next-line no-restricted-globals
+      const isExecuted = confirm("¿Estas segura de eliminar el incidente?");
+      if (isExecuted) deleteIncident();
+    }
   };
+
+  useEffect(() => {
+    if (type === "visit" && state.visits.length > 0) {
+      setDisabled(false);
+    } else if (type === "incident" && state.incident) {
+      setDisabled(false);
+    } else {
+      setDisabled(true);
+    }
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [state]);
+  // Mejorar rendimiento
   return (
     <button
       className="Button bg-error text-white flex justify-center gap-3 items-center text-base disabled:opacity-70"
-      disabled={state.visits.length === 0}
+      disabled={disabled}
       onClick={handleClick}
     >
-      <span>LIMPIAR VISITAS</span>
+      {children}
       <FaTrash />
     </button>
   );
